% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deformetrics.R
\docType{package}
\name{deformetrics-package}
\alias{deformetrics-package}
\alias{deformetrics}
\title{R Functions for Shape Analysis}
\description{
Tools for geometric morphometric analysis.
}
\author{
Paolo Piras, Antonio Profico, Valerio Varano, Pasquale Raia, Alessio Veneziano
}
