% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{internal_functions}
\alias{internal_functions}
\title{internal_functions}
\usage{
array2list(array)
array2mat(array,inds,vars)
biharm.new(vec1,vec2)
centershapes(array)
centroids(array)
firstsfac(group)
helmert(p)
lastsfac(group)
lshift2(array,factor,CR=NULL,locs=NULL,CSinit=F,scale=F,reflect=F,reorder=F,sc=F)
makeTransparent(..., alpha=0.5)
manymultgr(y,x,group,steps=5000)
mopa(tt,yy,rot=c("mopa","opa"),CSinit = FALSE,volinit=FALSE,center=TRUE)
newmb(source,target)
opaloop2(GM,array,scale=F,reflect=F)
ordiwithshapes(mshape,scores,rotation,whichaxes=c(1,2))
plotptau5(objptau,links=NULL,projorig=T,whichaxes=c(1,2))
pwpermancova(y, pred1,pred2,nperm=999)
ratesbygroup(predictions,factor,indep,sss=F)
rep.row(x,n)
scaleshapes(array)
seqrate(preds,factor,vector=centroid.size(preds),sss=F)
showPC2(x,rotation,mshape)
sigm.new(mat1,mat2=NULL)
tpsgridpaolo(TT, YY)
}
\description{
Here are reported a collection of internal functions
}
\author{
Paolo Piras
}
