% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lshift2.r
\name{lshift2}
\alias{lshift2}
\title{lshift2}
\usage{
lshift2(array, factor, CR = NULL, locs = NULL, CSinit = F, scale = F,
  reflect = F, reorder = F, sc = F)
}
\arguments{
\item{array}{numeric: an array kxmxn of landmark coordinates}

\item{factor}{character vector: a factor with group affiliation}

\item{CR}{logical: if NULLL CR is estimated as the consensus of the array via Generalized Procrustes Analysis (GPA). Otherwise the specified CR is used.}

\item{reflect}{logical:  “reflect” argument in procSym() from package “Morpho”}

\item{reorder}{logical:  If TRUE the factor is reordered}

\item{sc}{logical: This is not supposed to be called by common user}

\item{locs=NULL:}{if NULL the local references are estimated via separate per-group GPAs. Oterwise the specified locs are used.}

\item{CSinit=F}{logical: if TRUE scaling at unit size is performed.}

\item{scale=F}{logical “scale” argument in procSym() from package “Morpho”}
}
\value{
transported list: a list of length 1 containing the array with the transported shapes.
}
\description{
This function performs the Euclidean Parallel Transport, i.e. the Linear Shift strategy, described in Piras et al (2016). It applies a series of deformations occurring in different groups to a Common Reference by properly managing rotations.
}
\examples{
\dontrun{ 
data(case1)
group<-factor(rep(1:5,each=21))
linksdors<-conslinks(21)
library(gdata)
lscase1<-lshift2(case1,group,CR=procSym(case1[,,firstsfac(group)],CSinit=F,reflect=F,scale=F,pcAlign=F)$mshape,locs=case1[,,firstsfac(group)])
plottraj(procSym(lscase1$transported,CSinit=T)$PCscores[,1:2],group,asp=1) #### very similar to Levi Civita
}  
@export 
}
\references{
Piras P., Teresi L., Traversetti L., Varano V, Gabriele S., Kotsakis T., Raia P., Puddu P.E., Scalici M. (2016). The conceptual framework of ontogenetic trajectories: Parallel Transport allows the recognition and visualization of pure deformation patterns. Evolution and Development 18: 182-200. doi: 10.1111/ede.12186
}
\author{
Paolo Piras
}
