% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmyarrays.r
\name{plotmyarrays}
\alias{plotmyarrays}
\title{plotmyarrays}
\usage{
plotmyarrays(x, l = c(1:dim(x)[1]), v = c(1:dim(x)[2]),
  ind = c(1:dim(x)[3]), group = NULL, links = NULL, xlim = range(x),
  ylim = range(x), col = NULL, txt = T, lwd = 1, pch = NULL,
  cex = NULL, asp = 1, cextext = 1, xlab = "", ylab = "", zlab = "",
  xaxt = "s", yaxt = "s")
}
\arguments{
\item{x}{numeric: a matrix kxm or an array kxmxn of landmark coordinates}

\item{l}{numeric: is a vector indicating the number of variables of 1st array dimension to plot (default=c(1:dim(x)[1]))}

\item{v}{numeric: is a vector indicating the number of variables of 2nd array dimension to plot (default=c(1:dim(x)[2]))}

\item{ind}{numeric: is a vector indicating which elements of 3rd array dimension should be plotted (default=c(1:dim(x)[3]))}

\item{group}{character vector: optional grouping variable; shapes will be colored according to group affiliation}

\item{links}{numeric list: an optional list of vectors (default = NULL) indicating the links between landmarks}

\item{xlim}{numeric vector: xlim argument in par()}

\item{ylim}{numeric vector: ylim argument in par()}

\item{col}{numeric: col argument in par()}

\item{txt}{logical: if TRUE the text of ordinal landmarks number is plotted}

\item{lwd}{numeric: lwd argument in par() for links}

\item{pch}{numeric: pch argument in par() for points in 2D}

\item{cex}{numeric: cex argument in par() for points}

\item{asp}{numeric: giving the aspect ratio y/x (default=1)}

\item{cextext}{numeric: size of the text}

\item{xlab}{character: xlab argument in par() (default="")}

\item{ylab}{character: ylab argument in par() (default="")}

\item{zlab}{character: zlab argument in par() (default="")}

\item{xaxt}{character: xaxt argument in par() (default="s")}

\item{yaxt}{character: yaxt argument in par() (default="s")}
}
\description{
This function plot one or more elements of an array of three diensional or two dimensional shapes
}
\author{
Paolo Piras
}
