% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottraj.r
\name{plottraj}
\alias{plottraj}
\title{plottraj}
\usage{
plottraj(matrix, factor, label = NULL, sizelabel = 0, col = NULL,
  alpha = NULL, pch = NULL, lty = NULL, cex = NULL, axlab = NULL,
  xlim = range(matrix[, 1]), ylim = range(matrix[, 2]), asp = abs((xlim[1]
  - xlim[2])/(ylim[1] - ylim[2])), reorder = T, title = NULL, add = F)
}
\arguments{
\item{matrix}{a matrix kx2 or kx3}

\item{factor}{character: variable factor that affiliates each row to a trajectory}

\item{label}{character vector: optional vector of names}

\item{sizelabel}{numeric: size of the label text}

\item{col}{numeric: col argument in par()}

\item{pch}{numeric: pch argument in par()}

\item{lty}{numeric: lty argument in par()}

\item{cex}{numeric: cex argument in par()}

\item{xlim}{numeric vector: xlim argument in par() (default=range(matrix[,1]))}

\item{ylim}{numeric vector: ylim argument in par() (default=range(matrix[,2]))}

\item{asp}{numeric: giving the aspect ratio y/x (default=abs((xlim[1]-xlim[2])/(ylim[1]-ylim[2])))}

\item{reorder}{logical: if TRUE group is reordered according the the order of levels appearance}

\item{title}{character: optional name for the title (default=NULL)}

\item{add}{logical: if TRUE the plot is added to existing plot}

\item{alpha=NULL}{transparency value}
}
\description{
This function plot trajectory of data in two or three dimensions in the order they appear in the input matrix
}
\examples{
\dontrun{ 
data(case1)
group<-factor(rep(1:5,each=21))
plottraj(procSym(case1,CSinit=T)$PCscores[,1:2],group,asp=1)######## a classic approach does not recover the cycle
dtcase1<-newdt(case1,group,CR=procSym(case1[,,firstsfac(group)],CSinit=F,reflect=F,scale=F,pcAlign=F)$mshape,locs=case1[,,firstsfac(group)],tolrot=10)
plottraj(procSym(dtcase1,CSinit=T)$PCscores[,1:2],group,asp=1) ####  after the parallel transport it is correctly recovered
}
}
\author{
Paolo Piras
}
