% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt2dvar.r
\name{pt2dvar}
\alias{pt2dvar}
\title{pt2dvar}
\usage{
pt2dvar(array, group, doopa = F, tol = 1e-06, CR = NULL, locs = NULL,
  sss = T)
}
\arguments{
\item{array}{array: a kx2xn array of shapes}

\item{group}{character:    group structure}

\item{doopa}{logical: if TRUE Ordinary Procrustes Analysis is performed on the workhorse function. It is not supposed to be set at TRUE as OPAs are already performed outside that step.}

\item{tol}{numeric: tolerance parameter for rotation}

\item{CR}{matrix:common Reference to which intra-group deformations are applied}

\item{locs}{array:    Local references (one for each level of group) from which intra-group deformations are computed}

\item{sss}{logical: if TRUE analyses are performed on the Size and Shape Space. It is not supposed to be set at FALSE}
}
\value{
final list: a list with transported shapes
}
\description{
This function performs the Levi Civita Riemannian Parallel Transport according to explicit formulation presented in Varano et al (accepted) [Varano V, Gabriele S,  Teresi L, Dryden I L,  Puddu P E, Torromeo C,  Piras P. Accepted. The TPS Direct Transport: a new method for transporting deformations in the Size-and-shape Space. International Journal of Computer Vision].
}
\examples{
\dontrun{ 
data(case1)
groupsimul<-factor(rep(1:5,each=21))
plottraj(procSym(case1,CSinit=T)$PCscores[,1:2],group,asp=1)######## a classic approach does not recover the cycle
ptcase1<-pt2dvar(case1,groupsimul,CR=procSym(case1[,,firstsfac(groupsimul)],CSinit=F,reflect=F,scale=F,pcAlign=F)$mshape,locs=case1[,,firstsfac(groupsimul)])
plottraj(procSym(ptcase1$final,CSinit=T)$PCscores[,1:2],groupsimul,asp=1) ####  with this kind Parallel Transport the cycle is not perfectly recovered as with the new approach of Direct Transport (Varano et al [accepted]). 
}
}
\author{
Paolo Piras
}
