% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptau6.r
\name{ptau6}
\alias{ptau6}
\title{ptau6}
\usage{
ptau6(array, factor, CSinit = T, sepure = F, polyn = 1)
}
\arguments{
\item{array}{numeric: an array kxmxn of landmark coordinates}

\item{factor}{character: variable factor that affiliates shapes to group levels}

\item{CSinit}{logical: if TRUE shapes are scaled at unit size (default=TRUE)}

\item{polyn}{numeric: default=1 the degree of regression}

\item{sepur}{logical: if TRUE separate per-group multivariate regression between shape and size are performed on shapes aligned after separate GPAs  (default=FALSE)}
}
\description{
This function perform the Euclidean Parallel transport for allometric investigations using the Linear Shift startegy described in Piras et al (2016). The Linear Shift is performed on MANCOVA predictions and original data are projected on PCA space computed on them.
}
\examples{
\dontrun{ 
library(Morpho)
library(gdata)
library(rgl)
data(group)
data(my2d)
mysel<-group\%in\%c("Macroscelides_proboscideus","Petrodromus_tetradactylus","Elephantulus_rozeti","Elephantulus_edwardii")
linksdors<-list(c(1,2),c(37,7),c(12,4),c(27,28),c(25,21),c(38,40),c(9,10),c(2,3),c(3,4),c(1,7),c(1,6),c(3,5),c(6,40),c(5,9),c(40,8),c(8,9),c(1,7),c(7,6),c(3,4),c(4,5),c(39,38),c(38,35),c(35,37),c(37,39),c(35,34),c(34,33),c(33,32),c(32,31),c(31,30),c(30,29),c(29,37),c(37,36),c(36,29),c(28,31),c(28,30),c(13,10),c(10,11),c(11,12),c(12,13),c(13,14),c(14,16),c(16,17),c(17,20),c(20,19),c(19,18),c(18,12),c(18,15),c(15,12),c(21,19),c(21,20),c(24,25),c(25,26),c(26,27),c(27,24),c(26,24),c(24,23),c(23,22),c(22,8),c(8,2))       
dors4<-my2d[,,mysel]
factordors4<-drop.levels(group[mysel],reorder=T)
factordors4<-factor(factordors4,levels=unique(factordors4))
adors4<-procSym(dors4,scale=F,pcAlign=F,reflect=F)
mypredictbook<-read.inn(predict(lm(array2mat(adors4$orpdata,105,80)~poly(adors4$size,1,raw=T)*factordors4)),40,2)
procbook<-procSym(mypredictbook,pcAlign=F)
# linear shift in the shape space
plot(procbook$PCscores[,1:2],pch=as.numeric(factordors4),asp=1)
objptau<-ptau6(dors4,factordors4,CSinit=T)
#  morphological apprciation is dramatically different!
plotptau5(objptau,linksdors,pch=as.numeric(objptau$factorord),col=rep(1,length(objptau$factorord)),round(max(centroid.size(objptau$arrayord)),digits=0)/2,shiftnegy=2,mag=2,subplotdim=1)
## End(Not run)
}

}
\author{
Paolo Piras
}
