% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.inn.r
\name{read.inn}
\alias{read.inn}
\title{read.inn}
\usage{
read.inn(matrix, k, m, type = 0)
}
\arguments{
\item{matrix}{numeric: an array kxmxn of landmark coordinates}

\item{k}{numeric: number of landmarks}

\item{mtype}{numeric: if 0, if 1}
}
\description{
This function convert a matrix in an array
}
\author{
Paolo Piras
}
