% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoresequencearray.r
\name{scoresequencearray}
\alias{scoresequencearray}
\title{scoresequencearray}
\usage{
scoresequencearray(mshape, from, to, pcn, rotation, mag = 1, frames = 15,
  type = 1)
}
\arguments{
\item{mshape}{numeric: a kx2 matrix which will be deformed according to "from" and "to" values and to "rotation" eigenvectors.}

\item{from}{numeric: a starting value for the computation od the sequence of shapes}

\item{to}{numeric: the final value for the computation od the sequence of shapes}

\item{pcn}{numeric: column number of "rotation" argument to be used to deform "mshape"}

\item{rotation}{numeric: rotation}

\item{mag}{numeric: magnification (default=1)}

\item{frames}{numeric: number of frames to be saved (default=15)}
}
\value{
myarray array: an kxmxn array of predited shapes
}
\description{
This function computes a sequence of shapes computed by deformed a reference shape ("mshape" argument) according to the eigenvectors specified in "rotation" argument. The sequence is computed in the interval between "from" and "to" argument. Usually the "rotation" argument comes from a PCA but any ordination method could be used.
}
\author{
Paolo Piras
}
