% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublinks.r
\name{sublinks}
\alias{sublinks}
\title{sublinks
This function creates a new links file starting from a links file and a vector of lansmarks selection}
\usage{
sublinks(links, subvec)
}
\arguments{
\item{links}{list: A list of vectors each with two elements indicating the landmarks connected by a link}

\item{subvec}{list:  A vector ordered in ascending order indicating the subselection of landmarks for which the new link file should be created}
}
\value{
res list: A list with new links
}
\description{
sublinks
This function creates a new links file starting from a links file and a vector of lansmarks selection
}
\examples{
\\ dont run{
data(linksentire)
data(pri3d)
newlinks<-sublinks(linksentire,c(1:17))
plotmyarrays(pri3d[1:17,,1],links=newlinks)
}
}
\author{
Paolo Piras
}
