% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{internal_functions}
\alias{internal_functions}
\title{internal_functions}
\usage{
areasip()
array2list()
array2mat()
biharm.new()
centershapes()
centroids()
conslinks()
diffonmesh()
firstsfac()
helmert()
heat2d()
heat3d()
lastsfac()
list2matrix()
makeTransparent()
manymultgr()
mopa()
mycca()
newmb()
opaloop()
opaloop2()
ordiwithshapes()
plotdefo3d()
plotptau5()
plotsurf()
plsgen()
posfac()
pt2dvaruno()
pwpermancova()
ratesbygroup()
reparray()
rep.row()
scaleshapes()
seqrate()
serpred()
showPC2()
sigm.new()
tessell3d()
tpsgridpaolo()
}
\description{
Here are reported a collection of internal functions
}
\author{
Paolo Piras
}
