% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{internal_functions}
\alias{internal_functions}
\title{internal_functions}
\usage{
areasip(matrix,links=NULL,PB=NA,PA=NA,S=NA,SB=NA,H=NA,V=3,a=NULL,q=NULL,Y=FALSE,j=FALSE,D=FALSE,St=Inf,Q=TRUE,graph=T,extpol=F)
array2list(array)
array2mat(array,inds,vars)
biharm.new(vec1,vec2)
centershapes(array)
centroids(array)
diffonmesh()
firstsfac(group)
helmert(p)
heat2d(init,fin...)
lastsfac(group)
list2matrix(mylist)
makeTransparent(..., alpha=0.5)
manymultgr(y,x,group,steps=5000)
mopa(tt,yy,rot=c("mopa","opa"),CSinit = FALSE,volinit=FALSE,center=TRUE)
mycca()
newmb(source,target)
opaloop()
opaloop2(GM,array,scale=F,reflect=F)
ordiwithshapes(mshape,scores,rotation,whichaxes=c(1,2))
plotptau5(objptau,links=NULL,projorig=T,whichaxes=c(1,2))
posfac()
pt2dvaruno()
pwpermancova(y, pred1,pred2,nperm=999)
ratesbygroup(predictions,factor,indep,sss=F)
reparray()
rep.row(x,n)
scaleshapes(array)
seqrate(preds,factor,vector=centroid.size(preds),sss=F)
serpred()
showPC2(x,rotation,mshape)
sigm.new(mat1,mat2=NULL)
tpsgridpaolo(TT, YY)
}
\description{
Here are reported a collection of internal functions
}
\author{
Paolo Piras
}
