% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdt.r
\name{newdt}
\alias{newdt}
\title{newdt}
\usage{
newdt(array, group, CR = NULL, pole = CR, diffpole = F, locs = NULL,
  center = T, CSinit = F, tolrot = 1, tol = 1e-08, doopa = T,
  domopa = T)
}
\arguments{
\item{array}{numeric: an array kxmxn of landmark coordinates}

\item{group}{character vector: a factor with group affiliation}

\item{CR=NULL:}{if NULLL CR is estimated as the consensus of the array via Generalized Procrustes Analysis (GPA). Otherwise the specified CR is used.}

\item{pole=CR;}{by default the Riemannian pole coincides with CR.}

\item{diffpole=F}{logical: if TRUE the pole is not coincident with CR. This is not supposed to be used by common users.}

\item{locs=NULL:}{if NULL the local references ar estimated via separate per-group GPAs. Oterwise the specifiesed locs are used-}

\item{center=T}{logical: if TRUE shapes are centered.}

\item{CSinit=F}{logical: if TRUE caling at unit size is performed.}

\item{tolrot=1}{numeric: tolerance for rotation during GPA.}

\item{tol=1e-8}{numeric: tolerance for eigenvalue decomposition}

\item{doopa=T}{logical: if TRUE locs are aligned with CR via Ordinary Procrustes Analyisis (OPA)}

\item{domopa=T}{logical: if TRUE each shape within each group is aligned with its proper loc via Modified Ordinary Procrustes Analysis introduced in Varano et al (accepted)}
}
\value{
out matrix: a matrix containing the transported shapes.
}
\description{
This function performs the TPS Parallel Transport according to Varano et al (accepted). Deformation occurring within groups are estimate relatively to per-group local references ("locs" argument") and applied to a Common Reference ("CR").
}
\examples{
\dontrun{ 
data(case1)
group<-factor(rep(1:5,each=21))
plottraj(procSym(case1,CSinit=T)$PCscores[,1:2],group,asp=1)######## a classic approach does not recover the cycle
dtcase1<-newdt(case1,group,CR=procSym(case1[,,firstsfac(group)],CSinit=F,reflect=F,scale=F,pcAlign=F)$mshape,locs=case1[,,firstsfac(group)],tolrot=10)
}  
}
\author{
Paolo Piras
}
