% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmyarrays.r
\name{plotmyarrays}
\alias{plotmyarrays}
\title{plotmyarrays}
\usage{
plotmyarrays(x, l = c(1:dim(x)[1]), v = c(1:dim(x)[2]),
  ind = c(1:dim(x)[3]), group = NULL, links = NULL, xlim = range(x),
  ylim = range(x), col = NULL, txt = T, lwd = 1, pch = NULL,
  cex = NULL, asp = 1, cextext = 1, xlab = "", ylab = "", zlab = "",
  xaxt = "s", yaxt = "s")
}
\arguments{
\item{x}{numeric: a matrix kxm or an array kxmxn of landmark coordinates}

\item{l}{numeric: is a vector indicating the number of variables of 1st array dimension to plot (default=c(1:dim(x)[1]))}

\item{v}{numeric: is a vector indicating the number of variables of 2nd array dimension to plot (default=c(1:dim(x)[2]))}

\item{ind}{numeric: is a vector indicating which elements of 3rd array dimension should be plotted (default=c(1:dim(x)[3]))}

\item{group}{character vector: optional grouping variable; shapes will be colored according to group affiliation}

\item{links}{numeric list: an optional list of vectors (default = NULL) indicating the links between landmarks}

\item{xlim}{numeric vector: xlim argument in par()}

\item{ylim}{numeric vector: ylim argument in par()}

\item{col}{numeric: col argument in par()}

\item{txt}{logical: if TRUE the text of ordinal landmarks number is plotted}

\item{lwd}{numeric: lwd argument in par() for links}

\item{pch}{numeric: pch argument in par() for points in 2D}

\item{cex}{numeric: cex argument in par() for points}

\item{asp}{numeric: giving the aspect ratio y/x (default=1)}

\item{cextext}{numeric: size of the text}

\item{xlab}{character: xlab argument in par() (default="")}

\item{ylab}{character: ylab argument in par() (default="")}

\item{zlab}{character: zlab argument in par() (default="")}

\item{xaxt}{character: xaxt argument in par() (default="s")}

\item{yaxt}{character: yaxt argument in par() (default="s")}
}
\description{
This function plot one or more elements of an array of three diensional or two dimensional shapes
}
\examples{
\dontrun{  
data(my2d)
linksdors<-list(c(1,2),c(37,7),c(12,4),c(27,28),c(25,21),c(38,40),c(9,10),c(2,3),c(3,4),c(1,7),c(1,6),c(3,5),c(6,40),c(5,9),c(40,8),c(8,9),c(1,7),c(7,6),c(3,4),c(4,5),c(39,38),c(38,35),c(35,37),c(37,39),c(35,34),c(34,33),c(33,32),c(32,31),c(31,30),c(30,29),c(29,37),c(37,36),c(36,29),c(28,31),c(28,30),c(13,10),c(10,11),c(11,12),c(12,13),c(13,14),c(14,16),c(16,17),c(17,20),c(20,19),c(19,18),c(18,12),c(18,15),c(15,12),c(21,19),c(21,20),c(24,25),c(25,26),c(26,27),c(27,24),c(26,24),c(24,23),c(23,22),c(22,8),c(8,2))                                                                                                                                                           
my2d<-procSym(my2d,CSinit=T,scale=F)
pc1seq<-scoresequencearray(my2d$mshape,min(my2d$PCscores[,1]),max(my2d$PCscores[,1]),1,my2d$PCs)
par(mfrow=c(4,4))
for(i in 1:dim(pc1seq)[3]){
plotmyarrays(pc1seq[,,i],links=linksdors,cex=0,txt=F)}
}
}
  
\author{
Paolo Piras
}
